Big change to how the policy works, it now scans for the .foo VCS directories.
I realized a command driven version wasn't going to work well
* Windows users often use GUIs like TortiseSVN which have no command
* Configuration for out of path commands would be PER USER and PER VCS not
  per project.  It would be a pain in the ass
* It would be really slow as it would have to be done for every file
* Possibility of side effects

Now it just looks up the tree for the .foo VCS directory which covers
most VCS.  Perforce and SVK get an automatic pass for now.  This also
no longer requires the project to pre-configure the policy with the
type of VCS in use, but you can still do that if you want.

I also improved how it keeps it down to one check by telling PBP that
this policy applies to the whole document.

Still lacking real tests.